Simplify Gemfile and configuration for use with GitHub Pages and custom theme
Now using the `github-pages` gem to automatically use the  GitHub Pages dependency versions on local installs as well. See https://pages.github.com/versions/

Furthermore, the `remote_theme` plugin has been dropped in favor of simply using the `gem` syntax with `git:` to point to the remote theme hosted on GitHub, without the limitations that come with using the `jekyll-remote-theme` plugin, and without having the custom theme to be published to RubyGems.org.